+++ src/embed.js
@@ -3,6 +3,7 @@
 
     var baseUrl = win.__C6_URL_ROOT__ || '//portal.cinema6.com',
         appJs = win.__C6_APP_JS__ || '//lib.cinema6.com/c6embed/v1/app.min.js',
+        bools = ['preload'],
         config = (function(scripts) {
             var script = scripts[scripts.length - 1],
                 attributes = script.attributes,
@@ -23,6 +24,14 @@
                 }
 
                 result[prop] = value;
+            }
+
+            length = bools.length;
+
+            while (length--) {
+                attribute = bools[length];
+
+                result[attribute] = attribute in result;
             }
 
             result.script = script;
@@ -46,11 +55,12 @@
             requireCache: {},
             branding: {},
             gaAcctId: 'UA-44457821-2',
-            loadExperience: function(embed) {
+            loadExperience: function(embed, preload) {
                 var app = this.app || (this.app = document.createElement('script')),
                     head = document.getElementsByTagName('head')[0];
 
                 embed.load = true;
+                embed.preload = !!preload;
 
                 if (!app.parentNode) {
                     app.src = appJs;
@@ -76,6 +86,7 @@
             embed: div,
             splashDelegate: {},
             load: false,
+            preload: false,
             config: config
         };
 
@@ -92,6 +103,11 @@
         }
 
         return element;
+    }
+
+    function handleMouseenter() {
+        c6.loadExperience(settings, true);
+        splash.removeEventListener('mouseenter', handleMouseenter, false);
     }
 
     function splashOf(splashConfig) {
@@ -183,6 +199,12 @@
                     splash: baseUrl + '/collateral/experiences/' + config.exp + '/splash'
                 });
                 settings.splashDelegate = splashJS(c6, settings, splash);
+
+                if (config.preload) {
+                    c6.loadExperience(settings, true);
+                } else {
+                    splash.addEventListener('mouseenter', handleMouseenter, false);
+                }
             });
         });
     });